\name{ref_bkg_exome}
\alias{ref_bkg_exome}
\docType{data}
\title{Trinucleotide counts in the human exome}
\description{
Data frame with the human exome counts corresponding to the 96 single-nucleotide-substitution trinucleotide contexts (adapted from package \code{deconstructSigs} by Rosenthal et al.). It is used by the \code{sig.bkg.adj} argument in \code{\link{mutSigMapper}}.
}
\usage{data(ref_bkg_exome)}

\references{Rosenthal R et al.
\emph{DeconstructSigs: delineating mutational processes in single tumors distinguishes DNA repair deficiencies and patterns of carcinoma evolution}, Genome Biol. 17, 31 (2016).
}

\keyword{datasets}

