\name{ref_cosmic_v2}
\alias{ref_cosmic_v2}
\docType{data}
\title{COSMIC v2 (March 2015) compendium}
\description{
Compendium of 30 annotated mutational signatures represented by 96 single-nucleotide-substitution trinucleotide contexts based on the observed trinucleotide frequency of the human genome.
}
\usage{data(ref_cosmic_v2)}

\format{
A list that contains:
\itemize{
\item{\code{sig}: Data frame with substitution type, subtype, and signatures (columns) across 96 trinucleotide contexts (rows).}
\item{\code{desc}: Data frame with etiology annotations for each signature in the compendium.}
}
}
\references{
COSMIC (\emph{Catalogue Of Somatic Mutations In Cancer}), https://cancer.sanger.ac.uk/cosmic/signatures/index.tt
}

\keyword{datasets}

