\name{ref_cosmic_v3_exome}
\alias{ref_cosmic_v3_exome}
\docType{data}
\title{COSMIC v3 (May 2019) compendium (exome reference)}
\description{
Compendium of 65 annotated mutational signatures represented by 96 single-nucleotide-substitution trinucleotide contexts based on the observed trinucleotide frequency of the human exome.
}
\usage{data(ref_cosmic_v3_exome)}

\format{
A list that contains:
\itemize{
\item{\code{sig}: Data frame with substitution type, subtype, and signatures (columns) across 96 trinucleotide contexts (rows).}
\item{\code{desc}: Data frame with etiology annotations for each signature in the compendium.}
}
}
\references{
COSMIC (\emph{Catalogue Of Somatic Mutations In Cancer}), https://cancer.sanger.ac.uk/cosmic/signatures/index.tt
}

\keyword{datasets}

