\name{ref_mutagen53}
\alias{ref_mutagen53}
\docType{data}
\title{Mutagen compendium (Kucab et al, May 2019)}
\description{
Compendium of 53 annotated mutational signatures represented by 96 single-nucleotide-substitution trinucleotide contexts based on the observed trinucleotide frequency of the human genome.
}
\usage{data(ref_mutagen53)}

\format{
A list that contains:
\itemize{
\item{\code{sig}: Data frame with substitution type, subtype, and signatures (columns) across 96 trinucleotide contexts (rows).}
\item{\code{desc}: Data frame with treatment annotations for each signature in the compendium.}
}
}
\references{
Kucab JE, Zou X, Morganella S, ..., Arlt VM, Phillips DH, and Nik-Zainal S, 
\emph{A Compendium of Mutational Signatures of Environmental Agents}, Cell 177, 821-836.e16 (2019).
}

\keyword{datasets}

