\name{spectra}
\alias{spectra}
\docType{data}
\title{Example spectra dataset}
\description{
Data frame that contains the mutational counts of 9 normal human adult stem cell samples from 3 different tissues (columns) along the 96 single-nucleotide-substitution trinucleotide contexts (rows) based on whole genome sequencing.
}
\usage{data(spectra)}

\references{
Blokzijl F et al.
\emph{Tissue-specific mutation accumulation in human adult stem cells during life},
Nature 538, 260-264 (2016).

Blokzijl F et al. 
\emph{MutationalPatterns: comprehensive genome-wide analysis of mutational processes}, 
Genome Med. 10, 33 (2018).
}

\keyword{datasets}

